Always offset popup by width of dock
because in some situations the popup would be at the
width of the dock on the previous workspace. For
example with the dock on inteligently hide and
one workspace with a full screen app and one without
when switching back and fourth between workspaces the
popup would overlap the dock, then have an awkward gap.
This keeps the gap consistant.