(Un)dodge dock for some cases
Depending on use configuration the workspace picker may need to
move out of the way of the dock or not, on a display by display
case. This gives the vfunc_allocate for the workspace picker
access to if the dock shows on multiple displays or just the
primary. then we choose to dodge the dock if that is true/false
and depending on the side the dock is on.